+++ server/src/model/page.spec.ts
@@ -1,13 +1,6 @@
-import { readFileSync } from 'fs'
 import * as path from 'path'
-import I from 'immutable'
-import { Factory } from './model/factory'
-import { Opt, PathHelper } from './model/utils'
-import { Bundle } from './model/bundle'
-import { PageNode, PageValidationKind, UNTITLED_FILE } from './model/page'
-import { Fileish } from './model/fileish'
-
-const REPO_ROOT = path.join(__dirname, '..', '..')
+import { PageNode, PageValidationKind, UNTITLED_FILE } from './page'
+import { first, FS_PATH_HELPER, makeBundle } from './util.spec'
 
 describe('Page', () => {
   let page = null as unknown as PageNode
@@ -147,155 +140,3 @@ describe('Page validations', () => {
     expect(page.getValidationErrors().errors.map(e => e.message).toArray().sort()).toEqual([PageValidationKind.MALFORMED_UUID, PageValidationKind.MISSING_TARGET].sort())
   })
 })
-
-describe('Book validations', () => {
-  it.skip('Missing page', () => {})
-  it.skip('Duplicate chapter title', () => {})
-  it.skip('Duplicate page', () => {})
-})
-
-describe('Bundle validations', () => {
-  it.skip('Missing book', () => {})
-  it.skip('No books are defiend', () => {})
-})
-
-describe('Factory', () => {
-  it('instantiates a new object when the key does not exist', () => {
-    let counter = 0
-    const f = new Factory(() => ({ thing: counter++ }))
-    expect(f.getIfHas('key1')).toBeUndefined()
-    expect(f.get('key1').thing).toEqual(0)
-    expect(f.getIfHas('key1')).not.toBeUndefined()
-    expect(f.get('key1').thing).toEqual(0)
-
-    expect(f.getIfHas('key2')).toBeUndefined()
-    expect(f.get('key2').thing).toEqual(1)
-    expect(f.getIfHas('key2')).not.toBeUndefined()
-  })
-  it('removesByKeyPrefix works', () => {
-    const f = new Factory((x) => ({ foo: x, bar: 'dummy-object' }))
-    f.get('keyPrefix1')
-    f.get('keyPrefix2')
-    f.get('anotherprefix')
-    f.get('not_a_keyPrefix')
-
-    expect(f.all().size).toEqual(4)
-    const removed = f.removeByKeyPrefix('keyPrefix')
-    expect(removed.size).toEqual(2)
-    expect(f.all().size).toEqual(2)
-  })
-})
-
-describe('Happy path', () => {
-  let bundle = null as unknown as Bundle
-
-  beforeEach(() => {
-    bundle = makeBundle()
-    bundle.load(read(bundle.absPath))
-  })
-  it('loads the book bundle', () => {
-    expect(bundle.exists()).toBeTruthy()
-    expect(bundle.isLoaded()).toBeTruthy()
-    expect(bundle.books().size).toBe(1)
-  })
-  it('loads the Book', () => {
-    const book = first(bundle.books())
-    loadSuccess(book)
-  })
-  it('loads a Page', () => {
-    const book = loadSuccess(first(bundle.books()))
-    const page = first(book.pages())
-    loadSuccess(page)
-  })
-})
-
-describe('The abstract ancestor class', () => {
-  let previousNodeEnv: Opt<string>
-  class MyNode extends Fileish {
-    protected getValidationChecks() { return [] }
-  }
-  class MyXMLNode extends MyNode {
-    protected parseXML = (doc: Document) => {
-      throw new Error('I-always-throw-an-error')
-    }
-  }
-  beforeEach(() => { previousNodeEnv = process.env.NODE_ENV })
-  afterEach(() => { process.env.NODE_ENV = previousNodeEnv })
-  it('marks a missing file as loaded but not existing', () => {
-    const f = new MyNode(makeBundle(), FS_PATH_HELPER, '/to/nowhere/filename')
-    expect(f.isLoaded()).toBe(false)
-    expect(f.exists()).toBe(false)
-    f.load(undefined)
-    expect(f.isLoaded()).toBe(true)
-    expect(f.exists()).toBe(false)
-  })
-  it('marks a file as loaded if there is no parseXML method', () => {
-    const f = new MyNode(makeBundle(), FS_PATH_HELPER, '/to/nowhere/filename')
-    f.load('the contents of a beutiful sunset')
-    expect(f.exists()).toBe(true)
-  })
-  it('sends one nodesToLoad when the object has not been loaded yet', () => {
-    const f = new MyNode(makeBundle(), FS_PATH_HELPER, '/to/nowhere/filename')
-    const v = f.getValidationErrors()
-    expect(v.errors.size).toBe(0)
-    expect(v.nodesToLoad.size).toBe(1)
-    expect(first(v.nodesToLoad)).toBe(f)
-  })
-  it('sends zero validation errors when the file does not exist', () => {
-    const f = new MyNode(makeBundle(), FS_PATH_HELPER, '/to/nowhere/filename')
-    f.load(undefined)
-    const v = f.getValidationErrors()
-    expect(v.errors.size).toBe(0)
-    expect(v.nodesToLoad.size).toBe(0)
-  })
-  it('sends all parse errros as a diagnostic message in production (instead of throwing them)', () => {
-    process.env.NODE_ENV = 'production'
-    const f = new MyXMLNode(makeBundle(), FS_PATH_HELPER, '/to/nowhere/filename')
-    f.load('>invalid-xml')
-    expect(f.getValidationErrors().errors.size).toBe(1)
-    const err = first(f.getValidationErrors().errors)
-    expect(err.message).toBe('I-always-throw-an-error')
-  })
-})
-
-describe('Bugfixes', () => {
-  it('clears parse errors when the file parses correctly', () => {
-    const bundle = makeBundle()
-    ignoreConsoleWarnings(() => {
-      bundle.load('<invalid this-is-intentionally-invalid-XML content')
-    })
-    expect(bundle.getValidationErrors().errors.size).toBe(1)
-    loadSuccess(bundle)
-    expect(bundle.getValidationErrors().errors.size).toBe(0)
-  })
-})
-
-export const read = (filePath: string) => readFileSync(filePath, 'utf-8')
-
-export const FS_PATH_HELPER: PathHelper<string> = {
-  join: path.join,
-  dirname: path.dirname
-}
-
-export function first<T>(col: I.Set<T> | I.List<T>) {
-  const f = col.toArray()[0]
-  expect(f).toBeTruthy()
-  return f
-}
-
-export const makeBundle = () => new Bundle(FS_PATH_HELPER, REPO_ROOT)
-
-export function loadSuccess<T extends Fileish>(n: T) {
-  expect(n.isLoaded()).toBeFalsy()
-  n.load(read(n.absPath))
-  expect(n.isLoaded()).toBeTruthy()
-  expect(n.exists()).toBeTruthy()
-  expect(n.getValidationErrors().errors.size).toBe(0)
-  return n // for daisy-chaining
-}
-
-export function ignoreConsoleWarnings(fn: () => void) {
-  const warnSpy = jest.spyOn(console, 'warn').mockImplementation()
-  fn()
-  warnSpy.mockRestore()
-}
