+++ server/src/server.spec.ts
@@ -2,8 +2,8 @@ import assert from 'assert'
 import * as xpath from 'xpath-ts'
 import { Position } from 'vscode-languageserver'
 import { DOMParser, XMLSerializer } from 'xmldom'
-import { calculateElementPositions } from '../model/utils'
-import { fixDocument, padLeft } from '../fix-document-ids'
+import { calculateElementPositions } from './model/utils'
+import { fixDocument, padLeft } from './fix-document-ids'
 
 describe('calculateElementPositions', function () {
   it('should return start and end positions using siblings when available', () => {
